+++ failing_examples/src/bin/generics.rs
@@ -1,0 +1,8 @@
+#[derive(Reformation)]
+struct A<T>{
+    x: T
+}
+
+fn main(){
+
+}